+++ packages/react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/newnivers/web-packages/compare/@newnivers/react@0.0.5...@newnivers/react@0.0.6) (2023-08-02)
+
+
+### Bug Fixes
+
+* **hooks/useDelay:** 불필요한 상수 값 export 제외 ([#21](https://github.com/newnivers/web-packages/issues/21)) ([e3a8eb7](https://github.com/newnivers/web-packages/commit/e3a8eb7237f61bbb390156dae2f88d17d9ee9005))
+
+
+### Features
+
+* **hooks/useDelay:** useDelay hook 기능 구현 ([#21](https://github.com/newnivers/web-packages/issues/21)) ([05610e3](https://github.com/newnivers/web-packages/commit/05610e36f82d5ead53cb8a288d305d5b2f585813))
+
+
+
+
+
 ## [0.0.5](https://github.com/newnivers/web-packages/compare/@newnivers/react@0.0.4...@newnivers/react@0.0.5) (2023-08-01)